Bug Fix: Long notes were stopping at system breaks.
This was a major correction, involving several files.
timeObjects now have an alignmentX attribute.
voice.timeObjects is now a sequence of midiChords
and midiRests, terminated by the rightmost barline.
The runningMarker.timeObjects do _not_ include a
rightmost barline.
The first moment in a midiChord or midiRest now has
a systemIndex attribute instead of a chordStart or restStart
attribute.
Known Bug: Moments seem to be being sent twice!