GUI improvements on front page
Scores are no longer disabled if midiAccess is
not available. This means that the link to the
score's About file still appears. However, if the
score needs midiAccess, a text to that effect
appears below the link, and the Resident Soundfont
synth is disabled in the output devices select.
Added a scoreRequiresMidiAccess(scoreIndex)
function to make things easier to update in future.