Improved TrackWorker behaviour
Replaced "trkCompleted" message by
"trkStopped" message. TrackWorkers send
this when the trk they are sending is stopped prematurely. Keyboard1 then sends the
remaining silent messages (noteOffs and
controller messages) before sending the
messages for the following trk.
Also: The Keyboard1.resetChannel() function
has been replaced by a doAllSoundOff()
function that is called at the end of a live performance, and a doAllControllersOff()
function that is called at the beginning of a
performance.