Perform time speed option implemented!
Speed can now be changed at perform time by calling
sequence.setSpeed(speedtoSet).
The Assistant Performer currently just allows the speed
to be set globally, but it would be quite feasible to
implement a speed control (slider) that could be changed
on page 2 while the score is performing.
This implementation means that none of the
msPositionInScore or msDurationInScore properties ever
change, and that the code is considerably smaller/simpler
than in the old Assistant Performer.