Improvements to speed control display.
The control now has a green background and border,
and is visibly disabled when appropriate.
TODO in Controls.setSpeedControl:
Centre the speedControlDiv between the tracksControl
and the performanceButtons. If there is not enough
space, shift the performance buttons to the right.
(The speedControlDiv has a fixed width.)
Possibly change the internal range of the speed
slider to be its width in pixels.