Fixed the bugs in performance buttons functionality.
Working on a performance option in which the
running cursors are pushed manually.
Currently, a special cursor is set -- but this is wrong.
I need to have cursor=none, and move a small SVG
manually so that it will change systems automatically.
The SVG always moves from left to right, while the
cursor moves right to left on alternate systems.