Updated all files for eslint
Known Bug: Live performance with Keyboard1 is not yet
working correctly.
The file .eslintrc.json has been created in the MyWebsite project directory. This is used to define the rules for writing
javascript, and is checked whenever a .js file is saved,
because the Options/Text_Editor/Javascript/EsLint option
is turned on in Visual Studio.