Keyboard1.js and TrackWorker.js corrected.
I have finally understood that TrackWorker requires
neither the channelIndex nor the letSound internal variables!
The "holdAllStop" pedal option has been deleted from
TrackWorker.js since it is not mentioned in TrkOptions.js.
This means that it is NOT necessary to define a midiChannel
attribute on the outputTracks in Score.js!!