Keyboard1 speed enhancement
The base speed for both live (Keyboard1) and
normal (Sequence) performances is set using the global
speed control. This can be done both before and during a performance.
Keyboard1 can also use continuous controllers (e.g. the
pitch wheel) to change the speed relative to this base value.
New: The global speed control is reset to 100% when the
"Start" button is clicked. It can be set to some other value
before clicking the "Go" button. (In Keyboard1
performances click "Stop" if necessary.)