Updated to latest residentSf2Synth 28.08.2017
Going through the gree code with a fine toothcomb...
Got as far as soundFont.createNoteInfo() and noticed that
the synth takes no account of some generators stored
in the SoundFont (e.g. "pan"). The soundfont is stereo,
but the synth only seems to use the first instrumentBag,
so is mono. Investigate further...