Simple Bug Fix: volHold was being set wrongly.
Tested in Pianola Music, Study 1 and in all presets on
the WebMIDISynthHostTestSite.
All presets are now working, though there were
disturbances in the French Horn looping the first
time I tested it. Second time, everything was smooth.
The Arachno GrandPiano still sounds better on the
VirtualMIDISynth, so there's oviously more work to do.
But that's a job for the Web MIDI experts..