Completed revision of soundFont.js and soundFontParser.js
These have now constructed keyLayer objects for all the
presets in a loaded soundFont. The values of the keyLayer
attributes have been converted to units that will subsequently
be convenient (in SoundFontSynthNote.js)
SoundFontSynthNote.js has not yet been updated to reflect
the new keyLayer attribute names (which contain their units
of measurement). So this commit is not fully functional.