Major commit. Study 1 and Pianola Music running!
Realised the conversion of soundFont amounts to
values that can be used more conveniently/efficiently
at runtime. This incidentally:
1. isolates the parsing of the soundFont.
2. makes the conversion easier to debug, should there
be anything wrong.
I'm not entirely happy that everything is correct, since
the piano sound does not decay as nicely as it does
with the VirtualMidiSynth.
But the concepts seem mostly clear, and that's a big
step forward!
TODO: debug in the WebMIDISynthHost.