Bug Fix: Staff nameElems are always attached to the first voice.
There are only as many staff names as there are staves.
Discovered this bug when testing the performance of a version of the Pianola Music score having two voices per staff.
Note that staff.staffName is a string, but that
staff.voiceElems[0] (=staffName) is an SVG element (that has a style).