Cursor ctor now creates an .msPosInScoreSequence attribute.
This sequence goes backwards if regions are repeated in the region sequence.
In other words, the msPosInScore values are in performed sequence, but they are not unique.
The Cursor ctor is complete, but has not yet been tested.
TODO: Create demo regionDefs and regionSequence elements in the Study 2 score, and write the code that reads them in Score.js.