Defined Moment.recordingData(), cleaned up TrackRecording.js
Moment.recordingData() returns a timestamped clone of the moment's
messages. Its important to return cloned messages, since the containing
midiObject is about to be reset.
The public TrackRecording interface now just contains:
    moments[] -- the recorded moments ( {timestamp, messages[]} )
    addMoment(moment) // used in score playback
    addMessage(message, timestamp) // used in live performances
Added Track.setToFirstRegion(). This is called at the end of a
performance to reset the performance start state.