Last commit before the summer break.
Demo regions perform correctly in Study 2.
The current state of the midi performance controls is now added
to the first moment in each region in each track, so that repeated
regions perform identically.
Known issues (to be tackled after the summer break):
1. The new cursor makes the systemIndex argument to
Controls.reportMsPos(...) redundant, so it should be removed. This
has deep consequences:
    i)   MidiChord and MidiRest constructors only need one argument.
    ii)  Moment does not need a systemIndex attribute.
    iii) Sequence does not need systemIndexToReport.
    iv) ???
2. the RunningMarker class has been made redundant by the new
Cursor class, so should be deleted completely. This may have
consequences for the rest of the code. Try simply deleting it, to see
where that leads.
3. I think I should probably completely reprogram TimePointer in terms
of the new Cursor, since RunningMarker is going to be deleted.
4. The Score code should be made to analyse regions defined in the
score file, not just using the temporary getRegionDefs() and
getRegionSequence() as at present.
5. I think I ought to go the whole hog, and reprogram *everything* in
Typescript. That ought to pay off eventually.