Added an allSoundOff message to those sent to each track at the start of each Region.
Note that:
1. Moritz now writes the <score:scoreData> element in scores
that have defined their regions.
2. The AssistantPerformer remains responsible for maintaining the
state of the controls sent at the beginning of each region.
I tried to program Moritz to write the controls directly in the scores'
Moments, but this turned out to be extremely difficult: Moments
dont exist in Moritz until they are written into a score. This relates
particularly to "continuously changing" envelopes, which Moritz
defines using a form of shorthand...