Completed the setting of start- and end-Markers in the presence of regions.
Has been extensively tested, and there are no known bugs.
TODO: add the active regionNames to the start- and endMarkers, starting with the existing stub functions in those classes.
Possibly also add the currently performing region to the cursor (needs a bit more programming, but should not be too difficult).