Removed redundant staff.isVisible attribute. (Staves are always visible.)
It has not been possible for Moritz to write invisible staves for some time.
Moritz has now also been finally updated for this, and the scores have been recompiled. They are, however logically identical to the older versions. (All the date stamps are different, and in Tombeau1 the padding for the endRegionTextInfoFrames has been improved.)