Bug Fix: Conductor correction. There are still other known bugs...
The conductor tracks msPositionInScore, not msPositionInPerformance!
Removed the trackMarker factor in the calculation of xFactor in conductor.doConducting(...), and refactored _msPositionInPerformance to _msPositionInScore.
Known Bug #1: Why is the cursor getting ahead of the performance? Does this only happen in fast performances? Is the Sequence queuing future messages? Is there a way to stop that?
Known Bug #2: Incorrect restarting after a conducted performance. Some variabes seem not to be being reinitialized properly.