Added reportTickOverload callback to Controls.js and Sequence.js
This callback (defined in Controls.js) is called if sequence.tick() can't keep up with the speed of a performance, so that moments having different msPositionInScore have had to be sent "synchronously" (in a tight loop).
The callback currently just calls console.warn(message), but this is going to change. TODO: display a temporary amber warning disc at the top of the TimeMarker.
sequence.tick() currently reports only the number of moments that
1. have been sent synchronously during the overload _and_
2. have timestamps that differ by more than 16ms.
Moments that are sent synchronously, but have timestamps that are less than 16ms apart, are not reported (See Sequence.js).