Implemented two conducting modes: conductTimer and conductCreep.
These modes can be used consecutively in the same performance,
When the start conducting button is clicked, and the cursor moves into the score area, conducting begins in the conductTimer mode. This is the mode in which mousemoveEvent.clientX controls the rate at which conductor.now() changes with respect to performance.now().
On mousedown, the mode switches to the conductCreep mode in which elapsed time is proportional to the distance travelled by mousemove events (independently of performance.now()). This is like pressing a pause button, except that conductor.now() changes when the mouse moves.
On mouseup, the mode switches back seamlessly to conductTimer (the default).