Replaced cursor.msPosDataMap by cursor.msPosDataArray and improved the creep algorithm.
Replacing the map by an array simplifies the code, and may even make it faster. The array contains msPosData objects that now have msPositionInScore attributes, and it is sorted with those attributes in ascending order.
The complete msPosData object attributes are:
  .msPositionInScore
  .alignmentX
  .yCoordinates
  .pixelsPerMs
The creep algorithm now updates the TimeMarker alignment continuously in the final 3 pixels before each alignment in the score. This means that the grey cursor no longer jumps ahead of the TimeMarker when creeping slowly.