Temp commit: (working on Conductor...) Known Bug: timer sometimes not stopped correctly when peformance is stopped.
ConductCreep mode now works with event.movementX and event.movementY attributes, not with event.clientX -- which is used only by conductTimer mode.
So conductCreep mode no longer has to keep track of _prevX, which is now set in switchToConductTimer().
SetIntervalHandles are pushed on a stack, so as to ensure that they all get cleared under all circumstances.