SoundFont(s) are now loaded in parallel, using Promises.
Added a SoundFont.js file (an es6 Module) containing a readSoundFontFileAsArrayBuffer() function and the SoundFont constructor that processes the returned ArrayBuffer.
This SoundFont Module contains code adapted from ResidentSf2Synth/soundFont.js (which has not changed). The Module's code is used exclusively by the AssistantPerformer.
Added MIDI patch names to the Constants.js module.
Tested and working for the GrandPiano soundFont, but there is a known bug: The Study 2 soundFont does not load properly. This could be a bug in either the file or the parser.