Cursor.js now takes account of disabled tracks.
Removed CursorBase class completely. Moved code from Cursor ctor to Cursor.set(...)
Account is taken of the special case in which there is no midiObject left at the left side of a system.
Check the following: I think the cursor is not updating correctly when the conductor is moving at a speed other than 100%, or when the conductor's speed has been changed while performing.