Enhancements to startMarker and endMarker behaviour.
These now position themselves on barlines rather than the following chord or rest.
An attempt to set an endMarker on the first barline in a system now displays an alert telling the user to set it at the end of the previous system.
Symmetrically, an attempt to set a startMarker on the last barline in a system now displays an alert telling the user to set it at the beginning of the following system.