Added setTrackPerformanceObjects() function in Score.js...
This is the beginning of a major revision of the whole code base. It creates a flat list of performed objects in each track, so that Sequence.js does not have to deal with regions at all.
This should lead to much simplified code down the line: For example, I should be able to delete RegionLinks altogether! This will, however mean inspecting the existing code very carefully, so see what is redundant and what not.