Undersized moments are no longer flagged using permanent blue markers...
In this version of the AP, each moment has a single timestamp and a list of messages that are sent "simultaneously" to the synth. The time at which the messages are actually sent (their timestamp) depends on the local timer, but that does not affect their "simultaneity".
As far as the performance logic is concerned, the smallest logical moment duration (the duration set in the score) is 1 millisecond. In practice, the time between two performed moments can be any floating point value greater than 0.
Every message in a saved (.json) recording has a timestamp (relative to the start of the recording) that is rounded to the nearest integer millisecond. This means that moments that were nominally 1ms apart in the score can appear to be synchronous in the recording.