Revised system.barlines so that marker setting now works correctly...
The system.barlinesPerInterpretation attribute has been changed to system.barlines. This is a simple array of barlines, independent of the interpretation value. Barlines are managed using their .alignment attribute. They have have an .msPosInScore attribute but no msPosInPerf attribute.
Revised the code for setting the startMarker and endMarker to account for the above changes. This has been extensively tested in the Pianola Music score, and seems to be working correctly.